+++ lib/src/core/plugins.dart
@@ -231,7 +231,8 @@ class ZegoPrebuiltPlugins {
 
         /// not need logout
         // await ZegoUIKit().getSignalingPlugin().logout();
-        await ZegoUIKit().getSignalingPlugin().uninit();
+        /// not need destroy signaling sdk
+        await ZegoUIKit().getSignalingPlugin().uninit(forceDestroy: false);
       }
     }